+++ src/timestamp.rs
@@ -1,3 +1,3 @@
 /// Timestamp used to clear tracking information.
 #[derive(Clone, Copy)]
-pub struct TrackingTimestamp(pub(crate) u32);
+pub struct TrackingTimestamp(pub(crate) u64);
